ai_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-genai>=0.3.0
python-multipart>=0.0.9
